Added name to object and improved attributes in mesh.
All objects now have a name, which will improve debugging. Attributes
for the mesh have been improved to include normalization, as well as
deletion. Attributes can also change type. All attributes must be the
same length as vertices ensuring valid data. However, this makes a
dependency where the vertices must be set and finalized before an other
attribute is added.

+++ GameEngine/GameEngine/src/graphics/ShaderParser.cpp
@@ -146,7 +146,7 @@ namespace engine { namespace graphics {
 
   void ShaderParser::SetCurrentShaderType(ShaderType::Type _type)
   {
-    if (_type < 0 || _type >= m_sources.size())
+    if (_type < 0 || static_cast<size_t>(_type) >= m_sources.size())
     {
       m_currentType = ShaderType::None;
       m_currentSource = nullptr;
